# openapi.yaml
openapi: 3.1.0
info:
  title: İkiteknik Eğitim Operasyon Platformu API
  version: 1.2.0
paths:
  /api/admin/authorization/catalog:
    get:
      operationId: getAuthorizationCatalog
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Available authorities and role presets
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthorizationCatalog'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/admin/users:
    get:
      operationId: listManagedUsers
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Search'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200':
          description: Users ordered by email
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedUserPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/admin/users/{userId}/authorities:
    put:
      operationId: replaceUserAuthorities
      security: [{bearerAuth: []}]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReplaceAuthoritiesRequest'}
      responses:
        '200':
          description: Updated user authorities
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedUser'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Search:
      name: search
      in: query
      schema: {type: string, maxLength: 100, default: ''}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Forbidden:
      description: Authenticated user lacks the required authority
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    NotFound:
      description: Referenced resource was not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Conflict:
      description: Unique code or resource state conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    AuthorizationCatalog:
      type: object
      required: [authorities, roles]
      properties:
        authorities: {type: array, items: {type: string}, uniqueItems: true}
        roles:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
            uniqueItems: true
    ReplaceAuthoritiesRequest:
      type: object
      additionalProperties: false
      required: [authorities]
      properties:
        authorities:
          type: array
          maxItems: 50
          uniqueItems: true
          items: {type: string}
    ManagedUser:
      type: object
      required: [id, email, enabled, authorities]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        enabled: {type: boolean}
        authorities: {type: array, items: {type: string}, uniqueItems: true}
    ManagedUserPage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [content]
          properties:
            content: {type: array, items: {$ref: '#/components/schemas/ManagedUser'}}
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
        first: {type: boolean}
        last: {type: boolean}
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code: {type: string}
        message: {type: string}
        timestamp: {type: string, format: date-time}